change to long unsigned form

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	unsigned int a = 1, b = 1, c, sum_even = 0;
+	long unsigned int a = 1, b = 1, c, sum_even = 0;
 
 	while (b < 4000000)
 	{
